requirements.txt
arrow==0.10.0
Django==1.11.1
django-admin-sortable==2.1
django-crispy-forms==1.6.1
django-enumfields==0.9.0
django-environ==0.4.3
django-formtools==2.0
enum34==1.1.6
psycopg2==2.7.1
python-dateutil==2.6.0
python-env==1.0.0
pytz==2017.2
six==1.10.0
stevedore==1.2.0
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
waitress==1.0.2
whitenoise==3.3.0
